Add ability to upload/push to a registry
This patch will allow users to upload to either a docker registry or
to a pulp crane server.  We want to support Sattelite in the future as
well.

Will read authorization data out of ~/.pulp/admin.conf if it exists, otherwise
it will prompt for Username and Password.

Most of pulp work contributed by Aaron Weitekamp