Add Atomic Verify
This command takes an image and checks the local images for newer version, if one
exists it tells the user to run it.  It also searches through all of the layers to
find any layers which might have a newer image, and then reports on them.

atomic verify test
Image 'test' contains a layer 'Apache:1.0:1' that is out of date image version 'Apache:1.0:2' is available.
You should rebuild the 'test' image useing docker build, it could be vulnerable.