Started adding support for satellite in addition to pulp.  Need to test it and integrate it into the rest of the system.  Waiting on Satellite server to test with.  Jenny Ramseyer
Added a push-to-satellite function, might have gotten the config stuff to work.  Jenny Ramseyer

refactoring and adding documentation

rechecked API calls.  More documentation

it builds.  Still not sure the config file parser is going to work--need to check section headings, but it builds.  Jenny Ramseyer

switched false to true in config.py

debugging the REST API calls

more bug fixing. May have fixed payload, need to check.  Have to decide what to do about publish

Fixed.  It should work now.  Need to fiddle around with create_repo more, but the rest should be good.  Thanks to David Davis for help with uploading an image.

debugging create_repo

fixing build error.

fixing some build errors.  Still can't push to Satellite or Pulp.  Not sure why, working on that.  Jenny Ramseyer

more bug fixing, cleaning up CLI

minor tweaks

fixed the push to pulp/satellite problem!  Thanks to Alec and Will for the help!  Now new error

Seems to be working now

Hooray it works.  Now we're getting connection errors, which I will try to debug. But I think it works.  Might have messed up an API call, which is why it won't connect?  Or maybe it is a server problem?  I will investigate.  --Jenny Ramseyer

committing before merge

Sending in this as a pull request.

satellite and pulp are now mutually exclusive arguments

Fixed the authentication problems in the code.

Debugging upload calls

Added debug mode

Figured out how to get content view id, org id, repo id, etc.  This relies on the user knowing their activation key, which I am not sure we can count on, but I think we can.  Currently it has the problem that we automatically take the first product id, which is wrong in many cases.  But, we can't let the user input the product id (they don't have it--the satellite web UI gives them a product id, but it is different from the internal product ids).  The other option would be to have the user input the name of the desired product, but that relies on them inputting it exactly as in the json output, which seems unlikely.  Also not very elegant.

Fixed the post request such that now, when we send a no-data post request, the _no data_ is seen as a json file, so the request goes through (this was to fix the problem where the server was rejecting our requests to get an upload id)

Note: you need to run this with sudo, which means that you need the admin file on your root directory (at least, in my case I needed to run it that way)

Turns out that, while the user doesn't see the activation key number on the website, it is in the URL of the activation key page

updated manpage

fixed some bugs, upload still doesn't work

Switched to multipart request

Fixed the uploading bug.  Now I have a different uploading bug, but it seems much more tractable.  Will update as progress continues

Now I get a 405 method not allowed error.  Not sure what's happening with that, but I will keep investigating

Now we require the user to pass in both the repo and the activation key.  This is standard for Satellite, and saves us from the horrible hacks I did before.  More robust code.  Problem is that both of these numbers are only available in the URL to their respective pages in the Satellite UI

Added to man page.  Also added better error handling for uploading

It works now.  Hooray

fixed it so we don't actually need to take a tar file.

Better error handling

fixed upgroup, haven't fixed bash yet

got bash completion working.

Moving push image to satellite/pulp to their respective files, instead of util

adding helpful configuration comment

making everything pep8 compatible

reverting back to old system

modified to make pep8 compatible

Signed-off-by: jramseye <jramseye@redhat.com>