atomic.sh: don't use --spc
There's no need to use --spc when running the rhel-tools image here
since it is already a privileged container. Also, using --spc will bring
in --rm, which is counter-productive here since the user would have to
start the container manually everytime before we can exec the command
inside of it.