atomic.sysconfig: tweak TOOLSIMG var
Fix the name so that it has the hyphen. Also add the user component
'rhel7/' so that the suggestion given by atomic during the redirect
makes sense (otherwise it suggests `atomic run rhel-tools`, which won't
work). This relies on the fact that `atomic run` will re-use the
existing container rather than creating a new one.