atomic.py: fix rpm-ostree arg handling
Calling atomic host commands would fail with the following error:

    TypeError: can only concatenate tuple (not "list") to tuple

This occurred because in _rpmostree(), args is a tuple, but aargs is a
list.

This patch simplifies the way extra args are handled by directly passing
a list to _rpmostree(), without unpacking. Also fixes the check to see
if the first argument given is '--'.