Imports: general reorganization
As per PEP8[1], there are basically three distinct groups of modules:

 1) standard library imports
 2) related third party imports
 3) local application/library specific imports

With that classification, we can give a general and consistent order
to the module imports.

The use of relative imports (the "dot" notation) for local modules
is indeed a matter of taste, but backed by PEP-0328[1] and already
used on Atomic/__init__.py.

[1] - https://www.python.org/dev/peps/pep-0008/#imports
[2] - https://www.python.org/dev/peps/pep-0328/

Signed-off-by: Cleber Rosa <crosa@redhat.com>