Move lengthy subfunctions to classes
Set a pattern where if an atomic subfunction
is lengthy, then we can break that out into its own class.
This reduces the overall size of the Atomic class and allows
new development and debug to be cleaner.

Instead of using the func function of argparser, we pass
the Class and Function that should be run as variables. For
example, the Atomic verify subfunction would now be:

    _class=Atomic.Atomic, func='verify'

The defined class is instantiated after parseargs is run.  We
then perform the set_args and call the function.