BUGFIX: Atomic diff leaks FD (BZ # 1285006)
Under the condition that you diff two images and one or both
of the images are not RPM based, no cleanup of the mounts and
dm devices was done. This was reported via BZ #1285006.

We now handle the exit and cleanup differently which counts
for this condition.

Also added a try/except for catching control-c, where if found
it will clean up the mounts and dm devices.