SCAN ERRORS: correct pruning and add unmount timeout
Correcting a bug observed by the openscap team where under
certain conditions, the deletion of a temporary image would
cause a traceback when noprune=False.

After discussing with dwalsh and vgoyal, we decided to revert
to the original PR implementation of a timeout as the lazy
umounts and deferred removals and deletion complicated matters.

We now have  simple timeout loop for unmounting by path to prevent
race failures between openscap, docker, the filesystem, and/or
dms.