atomic: harden shell invocations (1248038)
This addresses one of the two issues raised by bugzilla
1248038 where atomic is using shell=True on its subprocess
calls which exposes a security concern.

Two new functions were added to accomplish this.  The
first is a util.chk_call which is similar to util.subp
but it uses the python function check call instead of
Popen.

The second function is atomic.sub_env_strings.  Previously,
there was env variable substitution occuring with cmd_env and
the subprocess call.  That was one upside of the shell=True
usage.  Now, prior to calling util.chk_call, we pass the
intended command to sub_env_strings and use cmd_env to
perform the variable substitution.

We will address part 2 of the security concerns, which
is the handling of LABELS as strings, in a different
PR.