Removal of remote inspection from docker
We stopped carrying a patch in docker that allows for a RESTFUL
probe of remote registries.  This function is being replaced
by a new package called skopeo.

The remote_inspection def was removed from the Docker.Client
class extension in atomic.py and is replaced by a def called
util.skopeo which returns the same json-based results.