Atomic/scan.py: Add --rootfs option
Add the ability for atomic scan to deal with chroots on the host's
filesystem.  You can now pass a list of chroots to --rootfs and your
scanner will process them as if they were containers or images.  This
works with VMs too.

$ sudo virt-filesystems -d rhel7
/dev/sda1
/dev/sdb1
/dev/sdb2
/dev/rhel/home
/dev/rhel/root
$ sudo guestmount  -d rhel7 -m /dev/rhel/root --ro /tmp/rhel
$ sudo ./atomic scan --rootfs /tmp/rhel
...

Updated the man pages for atomic scan as well as it had fallen out of
date.

Small update that makes sure the host's os.environ is passed into
scanning container.

Closes: #371
Approved by: rhatdan