mount: add support for system containers
Use an xattr to store what drivers is used to manage the mountpoint.
This is important because OSTree containers are simply bind mount to
the rootfs of the container already checked out on the file system,
while images are checked out and must be "rm -rf" when unmounted.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #377
Approved by: jlebon