Change rules around declaring a default scanner.
Previously, with the understanding we would be shipping a single
generic scanner configuration file, would 1) install the config
file and we had preset a default scanner in atomic.conf.

With the realization that the scanner could be variable based
on the distribution, the generic file is not longer installed
by default. Therefore, we should probably not set a default.

The prior code failed if a default was not defined in
/etc/atomic.conf.  That code check has been altered and we now
adhere to the following rules:

* If no default is defined AND there is only one configured
 scanner, we assume the single scanner IS the default.

* If no default is defined AND there are multiple scanners, we
 error out and ask the user to define one.

* If there is a default defined, good to go!

Closes: #383
Approved by: rhatdan