Atomic/scan.py: Set default scanner
After reworking the logic in set_scanner(), we missed the
condition that if a default scanner is defined in atomic.conf,
it should be set as self.scanner unless otherwise overriden.

Also, moved the scan_list function to be executed prior to
set_scanner() as listing scanners should not require a user
to have a default scanner or set a specific scanner.

Closes: #408
Approved by: baude