Atomic/scan.py: Fail on Dead containers
When someone trys to scan a dead container (typically
with --containers or --all), we usually observe DM or
mount issues.  In turn, this usually causes Atomic to
traceback and fail.  Adding in a quick check for those
cases where we nicely fail and cite the Dead container
as a reason.

Closes: #446
Approved by: cgwalters