Retain persistent data from vulnerability scans
We now want to retain a summary of each results when
performing a vulnerability scan.  This persistent data
will be used for marking images or containers that are
vulnerable in both atomic and cockpit.  This marking of
the image has been implemented in atomic images.

There is a core function get_vulnerable_ids() which
can returns a list of uuids that are vulnerable.

Closes: #476
Approved by: rhatdan