iQIcBAABAgAGBQJXl4HLAAoJECY9bfLhY+Hqep0P/ih/gXvMYZnbUOkgSRi17Dka  rooLwowNuHbfjNn+r+eo2C6cl91jUDcyOFlSuk1YNiH2INRams25Ap4a5u7vPkmN  6FiXjXjKHbFYLW7WryevJ5Ac22lfW7bRUM5cb3d3VcwVjvwxJj2BR27e/AxNJZSY  V4h7DRl9948W+aFmbtkhBuAOIMttJ8jfwvMevSmNXnvWeW8WygfpPOrDVPg99By0  uNjL/OFR8Gl7zhTsVm5XMh7gxDPSSkYrObitG+6VL6Jtem9/dkiVtHshEQrF3S8W  XgZXowJ1+6HXl9H3bX2zGB+n44dtNm10jxxqZhxOQIkNlVrkbP5CieveRNwra8Fl  dJ76lLCA14OSLF8ZqrEalyQeqow7EexcW0ikdgHBqerL6rQ/5CRRMd/DpPfXKvkX  MVT+7tPl9DVD21PDa7kRHGWkrhT0yjNF+NKAlE2fS8nmnHKV6FE0rnvq1Jfyxkva  f4gXfv9WLS1cHHm7ba3ZaqSxv8zCn3QL4CtvzYKHrVXLhoQ2+TYmpVkiaiqffDKq  gYcOnR1J12wcqxCIHmk55Kr+/pwJWvW6W4fQo3WauY1EYiD8Vd+Bfeq9Yge/0s5v  6KX4Wzf35+lNQ8hJ3TRszmxKMPGa3GdDvIrcCKF165o1CU42lNpg7T6Ema99vEg+  z7D+ZZ38jXIPtR1zLnef  =JzHT  -----END PGP SIGNATURE-----
syscontainers: use temporarily CLI ostree for checkouts

This is required to process Docker whiteouts files in OSTree, which are
not enabled by this patch.

There is an issue in the way the RepoCheckoutOptions object is mapped by
glib, as the C struct is using bit fields that are not supported by the
introspection, so use the CLI version of OSTree for now.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #497
Approved by: jlebon