Atomic/sign.py|util.py Support sigstore from conf files
We now derive the proper sigstore from a series of YAML
configuration files in /etc/containers/registries.d.  These
configuration files can have: sigstore and sigstore-write as
keys and the values are a file path or http|s URL.

When signing an image, as long as the -d override is not used,
we use those values when writing local signatures.

Closes: #603
Approved by: rhatdan