syscontainers: store full image name
store the full image name, registry included, when pulling an image.

Some characters in the url are not usable for the name of an OSTree
ref, so escape them with '_'.  This change breaks old images already
pulled.

use "rm -rf /ostree/repo/refs/heads/ociimage" and pull the images
again.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #620
Approved by: cgwalters