Atomic/mount.py: shutil.rmtree input must be dir
The unmount function for system containers was failing
for images that had regular files in its / directory
because shutil.rmtree was being used.  The rmtree method
requires its input to a path of type dir.  Added a simple
condition to check the input path and use os.remove(path)
on files.

Closes: #759
Approved by: rhatdan