Use secondary disks for storage integration tests
Due to a couple of bugs related to loopback and sfdisk, we should
be using available secondary disks for testing atomic storage.  This
is part of the destructive testing we now do.

Also, configured vagrant to dispatch two disks so the tests can be done
there too.

Closes: #851
Approved by: baude