Add --storage to install|run
To have parity function with pull, we need to add --storage
to install and run.  This allows for overriding when atomic.type
is defined.  For example, if you pull an image that has
atomic.type=system but you really want to store it
on the docker backend.

Closes: #861
Approved by: baude