Atomic/scan.py: Return codes from scan should reflect scan results BZ 1387288
A user has asked for the return code from scans to reflect whether or not
all the scans were successful.  This was reported in https://bugzilla.redhat.com/show_bug.cgi?id=1387288

So now, a return code of 0 implies all scans were sucessful and the operation
completed normally.  A return code of 1 implies that not all scans were
successful OR the operation did not complete normally.

Closes: #919
Approved by: giuseppe