Running command inside running container fails
An exception was being thrown when trying to run a command inside
an already running container.  For example:

sudo atomic run foobar echo true

The problem was a lack of a setter for user_command in the container
object as well as some mishandled string <-> list conversions.

Closes: #938
Approved by: rhatdan