Fix registries.d/*.yaml config lookup
ri.assemble_fqdn() uses only the data submitted when creating the ri
instance, not the results of a lookup, which is necessary when the
registry is not specified by the user.  The results of a lookup are
available through the (cached) ri.fqdn property instead.

Without this, signing centos/httpd:latest would try to look up
configuration for /centos/httpd instead of docker.io/centos/httpd.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>

Closes: #937
Approved by: baude