Fix environment variable not being displayed
'atomic images info' on a system container image was not displaying
the environment variables in the container anymore. This was due to
get_backend_and_image_obj() function decomposing the img_obj.image
to not be the full image name. img_obj.name should be used instead.

Signed-off-by: Yu Qi Zhang <jerzhang@redhat.com>

Closes: #947
Approved by: giuseppe