rpm_host_install.py: new file
Refactor the RPM generation code for system containers in a new file so
that it can be shared with the Docker backend.

For fully supporting the same /exports structure we will need to add the
support for reading the manifest.json file as well, but since the Docker
backend doesn't use --set for settings of the container, preprocessing
files won't be very useful.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #949
Approved by: baude