util: Don't try to create directories at module import time
We shouldn't do anything at module import time, as that happens during builds,
where we don't want to touch the host system (and may not have privileges to do
so).

This required some refactoring of the locking code too, as simply
instantiating the decorator was creating files too.

Closes: https://github.com/projectatomic/atomic/issues/963

Closes: #966
Approved by: rhatdan