syscontainers: support rename of directories
Support rename of directories in the manifest file.

To do so we need to iterate each element of the path and see if there is
any substitution defined for the subpath itself.

Given a file `/usr/container/foo/bar` and a `renameFiles` like:

"renameFiles" : {
    "/usr/container" : "/usr/moved"
},

We iterate over these paths:

- /usr
- /usr/container <- this is renamed to  /usr/local/moved
- /usr/moved/foo
- /usr/moved/foo/bar <- final path used for the file

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #984
Approved by: ashcrow