syscontainers: use Skopeo copy to pull images
New versions of Skopeo support "ostree" as a destination for copy.  The
missing layers are written directly to the OSTree storage without any
additional handling from atomic.

Check if the used version of Skopeo has support for ostree and use
"skopeo copy" in this case.  In future we might drop completely the
other code path and assume ostree is always supported.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #1082
Approved by: baude