syscontainers: use repo under /var if on a different fs than /ostree/repo
If /ostree/repo is on a different file system than the checkout path,
use /var/lib/containers/atomic/.storage/ostree as the location for the
ostree repository.
Having the repository and the checkouts directory on the same file
system permits checkouts to be done through hard links instead of copied
files.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #1114
Approved by: giuseppe