syscontainers: introduce ALL_PROCESS_CAPABILITIES
it can be used by privileged containers to add all the capabilities
known to the host.  It prevents hard coding the list in the
config.json.template file so that the same image can be used on
different kernel versions where there can be different capabilities
supported.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #1119
Approved by: ashcrow